Add vmcExApp/
Make it a little bit easier for beginners to use this module
for a very first IOC with motion control.
On my system a binary is created here:
.../modules/motor/bin/darwin-x86/vmc (MacOs)
.../modules/motor/bin/windows-x64-mingw/vmc.exe (MINGW)

which is the one to be started as an example IOC.

Problem(s) that may needed to be solved:
The windows machine does not find the dlls, like asyn.dll
Copying them into .../modules/motor/bin/windows-x64-mingw/ allows to start
the binary.

More needs to be done, all macros need to be defined to get any record loaded.